// plugin/types.d.ts
export declare type premierepro = {
  AppPreference: AppPreferenceStatic
  AudioClipTrackItem: AudioClipTrackItemStatic
  AudioComponentChain: AudioComponentChainStatic
  AudioTrack: AudioTrackStatic
  CaptionTrack: CaptionTrackStatic
  ClipProjectItem: ClipProjectItemStatic
  EncoderManager: EncoderManagerStatic
  Exporter: ExporterStatic
  FolderItem: FolderItemStatic
  FrameRate: FrameRateStatic
  Guid: GuidStatic
  Keyframe: KeyframeStatic
  Marker: MarkerStatic
  Markers: MarkersStatic
  Metadata: MetadataStatic
  Project: ProjectStatic
  ProjectClosedEvent: ProjectClosedEventStatic
  ProjectEvent: ProjectEventStatic
  ProjectItem: ProjectItemStatic
  ProjectSettings: ProjectSettingsStatic
  ProjectUtils: ProjectUtilsStatic
  Properties: PropertiesStatic
  ScratchDiskSettings: ScratchDiskSettingsStatic
  SequenceEditor: SequenceEditorStatic
  SequenceUtils: SequenceUtilsStatic
  SourceMonitor: SourceMonitorStatic
  TickTime: TickTimeStatic
  TrackItemSelection: TrackItemSelectionStatic
  TransitionFactory: TransitionFactoryStatic
  Utils: UtilsStatic
  VideoClipTrackItem: VideoClipTrackItemStatic
  VideoComponentChain: VideoComponentChainStatic
  VideoFilterComponent: VideoFilterComponentStatic
  VideoFilterFactory: VideoFilterFactoryStatic
  VideoTrack: VideoTrackStatic
  VideoTransition: VideoTransitionStatic
}
// ...
export declare type AppPreferenceStatic = {
  setValue(): boolean	//Set backend preference using given list of property keys. The parameters are <key, value (number, boolean or string), persistence flag>
  getValue(preferenceKey: string): string	//Get preference value in native string form
  KEY_AUTO_PEAK_GENERATION: string	//Preference string key used to modify auto-peak generation settings
  KEY_IMPORT_WORKSPACE: string	//Preference string key used to modify import workspace settings
  KEY_SHOW_QUICKSTART_DIALOG: string	//Preference string key used to modify show quickstart dialog settings
  PROPERTY_PERSISTENT: number	//Property is persistent in backend and shared across cloud project.
  PROPERTY_NON_PERSISTENT: number	//Property is not persisted and will be cleared when the project closes.
}

export declare type AppPreference = {
}

export declare type Application = {
  version: string
}

export declare type AudioClipTrackItemStatic = {
  TRACKITEMTYPE_EMPTY: number	//Empty Track Item Type
  TRACKITEMTYPE_CLIP: number	//Clip Track Item Type
  TRACKITEMTYPE_TRANSITION: number	//Transition Track Item Type
  TRACKITEMTYPE_PREVIEW: number	//Previe Track Item Type
  TRACKITEMTYPE_FEEDBACK: number	//Feedback Track Item Type
}

export declare type AudioClipTrackItem = {
  getMatchName(): Promise<string>	//Returns the value of internal matchname for this trackItem
  getSpeed(): Promise<number>	//Returns the value of speed of the trackItem
  isAdjustmentLayer(): Promise<boolean>	//Returns true if the trackitem is an adjustment layer
  isSpeedReversed(): Promise<number>	//Returns true if the trackitem is reversed
  createMoveAction(tickTime: TickTime): Action	//Returns an action moves the inPoint of the track item to a new time, by shifting it by a number of seconds.
  getInPoint(): Promise<TickTime>	//Get timecode representing the inPoint of sequence.
  getOutPoint(): Promise<TickTime>	//Get timecode representing the outPoint of sequence.
  createSetInPointAction(tickTime: TickTime): Action	//Create SetInPointAction for sequence
  createSetOutPointAction(tickTime: TickTime): Action	//Create SetInPointAction for sequence
  getStartTime(): Promise<TickTime>	//Timecode representing the start of this track item relative to the sequence start.
  getEndTime(): Promise<TickTime>	//Timecode representing the end of this track item relative to the sequence start.
  getDuration(): Promise<TickTime>	//Timecode representing the duration of this track item relative to the sequence start.
  getType(): Promise<number>	//Index representing the type of this track item.
  isDisabled(): Promise<boolean>	//Returns true if rackitem is muted/disabled
  createSetDisabledAction(disabled: boolean): Action	//Returns an action that enables/disables the trackItem 
  getMediaType(): Promise<Guid>	//UUID representing the underlying media type of this track item
  getTrackIndex(): Promise<number>	//Index representing the track index of the track this track item belongs to
  getProjectItem(): Promise<ProjectItem>	//The project item for this track item.
  getComponentChain(): Promise<AudioComponentChain>	//Returns AudioComponentChain
}

export declare type AudioComponentChainStatic = {
}

export declare type AudioComponentChain = {
  createInsertComponentAction(component: object, componentInsertionIndex: number): Action	//Creates and returns an insert component action
  createAppendComponentAction(component: object): Action	//Creates and returns an append component action
  createRemoveComponentAction(component: object): Action	//Creates and returns an remove component action
  getComponentAtIndex(componentIndex: number): VideoComponentChain	//Returns the component at the given index
  getComponentCount(): number	//Gets the number of components in the component chain
}

export declare type AudioTrackStatic = {
  EVENT_TRACK_CHANGED: string	//Event Object for Track changed
  EVENT_TRACK_INFO_CHANGED: string	//Event Object for Track Info Changed
  EVENT_TRACK_LOCK_CHANGED: string	//Event Object for Track Lock Changed
}

export declare type AudioTrack = {
  setMute(mute: boolean): Promise<boolean>	//sets the mute state of the track to muted/unmuted
  getMediaType(): Promise<Guid>	//UUID representing the underlying media type of this track
  getIndex(): Promise<number>	//Index representing the track index of this track within the track group.
  isMuted(): Promise<boolean>	//Get mute state of the track
  getTrackItems(trackItemType: number, includeEmptyTrackItems: boolean): any	//Returns array of AudioClipTrackItem from the track item type
  name: string	//Get the name of the track
  id: number	//The ID of the track within the TrackGroup
}

export declare type CaptionTrackStatic = {
}

export declare type CaptionTrack = {
  setMute(mute: boolean): Promise<boolean>	//sets the mute state of the track to muted/unmuted
  getMediaType(): Promise<Guid>	//UUID representing the underlying media type of this track
  getIndex(): Promise<number>	//Index representing the track index of this track within the track group.
  isMuted(): Promise<boolean>	//Get mute state of the track
  getTrackItems(trackItemType: number, includeEmptyTrackItems: boolean): []	//This returns the track items of the specified media type from the given track
  name: string	//Get the name of the track
  id: number	//The ID of the track within the TrackGroup
}

export declare type ClipProjectItemStatic = {
  cast(projectItem: any): any
}
// ...
export declare type ExporterStatic = {
  exportSequenceFrame(sequence: object, time: TickTime, filename: string, filepath: string, width: number, height: number): Promise<boolean>	//Exports from a sequence. Supported formats are bmp, dpx, gif, jpg, exr, png, tga and tif
}

export declare type Exporter = {
}
// ...
export declare type Project = {
  getActiveSequence(): Promise<Sequence>	//Get the active sequence of the project
  setActiveSequence(sequence: Sequence): Promise<boolean>	//Set the active sequence of the project
  createSequence(name: string, presetPath?: string): Promise<Sequence>	//Create a new sequence with the default preset path - Parameter presetPath is deprecated, instead use createSequenceWithPresetPath()
  createSequenceFromMedia(name: string, clipProjectItems: [], targetBin?: ProjectItem): Promise<Sequence>	//Create a new sequence with a given name and medias
  getColorSettings(): Promise<ProjectColorSettings>	//Get project color settings object
  deleteSequence(sequence: Sequence): Promise<boolean>	//Delete a given sequence from the project
  getInsertionBin(): Promise<ProjectItem>	//Get current insertion bin
  openSequence(sequence: Sequence): Promise<boolean>	//Open a sequence and return true if successful.
  importSequences(projectPath: string, sequenceIds: []): Promise<boolean>
  importAEComps(aepPath: string, compNames: [], TargetBin?: ProjectItem): Promise<boolean>
  importAllAEComps(aepPath: string, TargetBin?: ProjectItem): Promise<boolean>
  importFiles(filePaths: [], suppressUI?: boolean, targetBin?: ProjectItem, asNumberedStills?: boolean): Promise<boolean>	//Import files in root/target bin of the project
  close(closeProjectOptions?: CloseProjectOptions): Promise<boolean>	//Close a project
  save(): Promise<boolean>	//Save the project
  saveAs(path: string): Promise<boolean>	//Save the project at the provided path
  getSequence(guid: Guid): Sequence	//Get sequence by id from the project
  getSequences(): Promise<[]>	//Get an array of all sequences in this project.
  getRootItem(): Promise<FolderItem>	//The root item of the project which contains all items of the project on the lowest level.
  pauseGrowing(pause: boolean): Promise<boolean>	//Pause growing of files instead swap the files
  executeTransaction(callback: any, undoString?: any): any	//Execute undoable transaction by passing compound action
  lockedAccess(callback: any): any	//Get a read/upgrade locked access to Project, project state will not change during the execution of callback function. Can call executeTransaction while having locked access.
  guid: Guid	//The unique identifier of the project.
  name: string	//The project name.
  path: string	//The absolute file path to the project file.
}
// ...
export declare type Sequence = {
  getSequenceVideoTimeDisplayFormat(): Promise<TimeDisplay>	//Get video time display format of this sequence
  getSequenceAudioTimeDisplayFormat(): Promise<TimeDisplay>	//Get audio time display format of this sequence
  getPlayerPosition(): Promise<TickTime>	//Get the player's current position
  setPlayerPosition(positionTime?: TickTime): Promise<boolean>	//Set the player's current position
  clearSelection(): Promise<boolean>	//Clears TrackItem Selection
  setSelection(trackItemSelection: TrackItemSelection): Promise<boolean>	//Updates sequence selection using the given track item selection.
  getVideoTrackCount(): Promise<number>	//Get video track count from this sequence
  getAudioTrackCount(): Promise<number>	//Get audio track count from this sequence
  getCaptionTrackCount(): Promise<number>	//Get caption track count from this sequence
  getVideoTrack(trackIndex: number): Promise<VideoTrack>	//Get video track from track index
  getAudioTrack(trackIndex: number): Promise<AudioTrack>	//Get audio track from track index
  getCaptionTrack(trackIndex: number): Promise<CaptionTrack>	//Get caption track from track index
  getSettings(): Promise<SequenceSettings>	//Get sequence settings object
  setSettings(sequenceSettings: SequenceSettings): Promise<boolean>	//Set sequence settings
  createCloneAction(): Action	//Creates an action to clone the given sequence
  createSubsequence(ignoreTrackTargeting?: boolean): Promise<Sequence>	//Returns a new sequence, which is a sub-sequence of the existing sequence
  isDoneAnalyzingForVideoEffects(): Promise<boolean>	//Returns whether or not the sequence is done analyzing for video effects
  getZeroPoint(): Promise<TickTime>	//Time representing the zero point of the sequence.
  getEndTime(): Promise<TickTime>	//Time representing the end of the sequence
  getInPoint(): Promise<TickTime>	//Get time representing the inPoint of sequence.
  getOutPoint(): Promise<TickTime>	//Get time representing the inPoint of sequence.
  createSetZeroPointAction(tickTime: TickTime): Action	//Create an action to set an InPoint for the sequence
  getProjectItem(): Promise<ProjectItem>	//Get the associated projectItem of the sequence.
  getSelection(): Promise<TrackItemSelection>	//Returns the current selection group of the sequence.
  getFrameSize(): Promise<RectF>	//Gets the size of the frame
  getTimebase(): Promise<string>	//Gets the time base of sequence
  guid: Guid	//The unique identifier of the sequence.
  name: string	//The sequence name.
}